Delete methods take int, but the ID model type is long.   At high IDs this overflows silently. Should be long throughout to match the model.

+++ Controllers/DepartmentsController.cs
@@ -42,7 +42,7 @@ namespace WebAPI.Controllers
         }
 
         [HttpDelete]
-        public HttpResponseMessage Delete(int id)
+        public HttpResponseMessage Delete(long id)
         {
             _departmentService.DeleteDepartment(id);
             return Request.CreateResponse(HttpStatusCode.OK);
